Moved match removing logic to only execute if the block request is approved.

+++ resources/block_requests.py
@@ -76,23 +76,25 @@ class   BlockRequestResource(Resource):
             block_request = BlockRequest.get(id=data.get("id", None))
 
             if block_request:
-                match = Match.check_match(block_request.reporter_id, block_request.reported_id)
 
-                if match["liked"] or match["matched"]:
-                    my_like = Match.get(matcher_id=block_request.reporter_id, matchee_id=block_request.reported_id)
-                    their_like = Match.get(matcher_id=block_request.reported_id, matchee_id=block_request.reporter_id)
+                if block_request.blocked:
+                    match = Match.check_match(block_request.reporter_id, block_request.reported_id)
                     
-                    if match["liked"] and match["matched"]:
-                        try:
-                            my_like.delete()
-                            their_like.delete()
-                        except Exception as e:
-                            return {"message" : str(e)}, 500
-                    elif match["liked"] and not match["matched"]:
-                        try:
-                            my_like.delete()
-                        except Exception as e:
-                            return {"message" : str(e)}, 500
+                    if match["liked"] or match["matched"]:
+                        my_like = Match.get(matcher_id=block_request.reporter_id, matchee_id=block_request.reported_id)
+                        their_like = Match.get(matcher_id=block_request.reported_id, matchee_id=block_request.reporter_id)
+                        
+                        if match["liked"] and match["matched"]:
+                            try:
+                                my_like.delete()
+                                their_like.delete()
+                            except Exception as e:
+                                return {"message" : str(e)}, 500
+                        elif match["liked"] and not match["matched"]:
+                            try:
+                                my_like.delete()
+                            except Exception as e:
+                                return {"message" : str(e)}, 500
 
                 block_request.reviewed = True
                 block_request.blocked = data["blocked"]
